refactor: extract base vaadin-details logic to reusable mixin
